requirements.txt
langchain==0.2.6
langchain_community==0.2.6
langchain_core==0.2.11
langchain_groq==0.1.6
langgraph==0.1.5
python-dotenv==1.0.1
streamlit==1.35.0
